+++ CHANGES.rst
@@ -4,7 +4,7 @@ claviger Changelog
 0.1.3 (unreleased)
 ==================
 
-- Nothing changed yet.
+- bugfix: actually set server name with ``name`` option.
 
 
 0.1.2 (2016-01-08)